added appropriate scaling for sampling
The coordinates used in the `warp` function of `skimage.transform` are
supposed to be integer coordinates or at least they have to scale with
the spatial shape of the inputs. However, this does not fit to the usual
domain we are using. Hence, the coordinates are from now on assumed to be
in [0,1]^d and the sampling procedure is adjusted accordingly to
incorporate the necessary scaling.